// typescript/flow-to-ts/Libraries/Components/View/ReactNativeViewViewConfigAndroid.d.ts
declare const ReactNativeViewViewConfigAndroid: {
    uiViewClassName: string;
    bubblingEventTypes: {
        topSelect: {
            phasedRegistrationNames: {
                bubbled: string;
                captured: string;
            };
        };
    };
    directEventTypes: {
        topClick: {
            registrationName: string;
        };
        topContentSizeChange: {
            registrationName: string;
        };
        topLoadingError: {
            registrationName: string;
        };
        topLoadingFinish: {
            registrationName: string;
        };
        topLoadingStart: {
            registrationName: string;
        };
        topMessage: {
            registrationName: string;
        };
        topMomentumScrollBegin: {
            registrationName: string;
        };
        topMomentumScrollEnd: {
            registrationName: string;
        };
        topScroll: {
            registrationName: string;
        };
        topScrollBeginDrag: {
            registrationName: string;
        };
        topScrollEndDrag: {
            registrationName: string;
        };
        topSelectionChange: {
            registrationName: string;
        };
    };
    validAttributes: {
        hasTVPreferredFocus: boolean;
        focusable: boolean;
        nativeBackgroundAndroid: boolean;
        nativeForegroundAndroid: boolean;
        nextFocusDown: boolean;
        nextFocusForward: boolean;
        nextFocusLeft: boolean;
        nextFocusRight: boolean;
        nextFocusUp: boolean;
    };
};
export default ReactNativeViewViewConfigAndroid;